+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 * fix create multiblock generator infinite rotation
 * add a node to deserialize a recipe to nbt
 * add an option for geckolib renderer to automatically `Schedule State Animation` without the node graph.
+* add multiblock debugger mod to tell why a multiblock is not assembling